hw06_pipeline_execution/pipeline_test.go
package hw06pipelineexecution

import (
	"fmt"
	"strconv"
	"testing"
	"time"

	"github.com/stretchr/testify/require"
)

const (
	sleepPerStage = time.Millisecond * 100
	// Погрешность набегает чуть больше, чем 50 мс, а именно 55, поэтому увеличиваю до 100.
	fault = sleepPerStage
)

func TestPipeline(t *testing.T) {
	// Stage generator
	g := func(_ string, f func(v interface{}) interface{}) Stage {
		return func(in In) Out {
			out := make(Bi)
			go func() {
				defer close(out)
				for v := range in {
					time.Sleep(sleepPerStage)
					out <- f(v)
					// Printing intermediate results
					fmt.Println(f(v))
				}
			}()
			return out
		}
	}

	stages := []Stage{
		g("Dummy", func(v interface{}) interface{} { return v }),
		g("Multiplier (* 2)", func(v interface{}) interface{} { return v.(int) * 2 }),
		g("Adder (+ 100)", func(v interface{}) interface{} { return v.(int) + 100 }),
		g("Stringifier", func(v interface{}) interface{} { return strconv.Itoa(v.(int)) }),
	}

	t.Run("simple case", func(t *testing.T) {
		in := make(Bi)
		data := []int{1, 2, 3, 4, 5}

		go func() {
			for _, v := range data {
				in <- v
			}
			close(in)
		}()

		result := make([]string, 0, 10)
		start := time.Now()
		for s := range ExecutePipeline(in, nil, stages...) {
			result = append(result, s.(string))
		}
		elapsed := time.Since(start)

		require.Equal(t, []string{"102", "104", "106", "108", "110"}, result)
		require.Less(t,
			int64(elapsed),
			// ~0.8s for processing 5 values in 4 stages (100ms every) concurrently
			int64(sleepPerStage)*int64(len(stages)+len(data)-1)+int64(fault))
	})

	t.Run("done case", func(t *testing.T) {
		in := make(Bi)
		done := make(Bi)
		data := []int{1, 2, 3, 4, 5}

		// Abort after 200ms
		abortDur := sleepPerStage * 2
		go func() {
			<-time.After(abortDur)
			close(done)
		}()

		go func() {
			for _, v := range data {
				in <- v
			}
			close(in)
		}()

		result := make([]string, 0, 10)
		start := time.Now()
		for s := range ExecutePipeline(in, done, stages...) {
			result = append(result, s.(string))
		}
		elapsed := time.Since(start)
		// require.Len(t, result, 0)
		// Не согласен. За 2 такта abortDur успевают просочиться в пайплайн 2 значения, и они доходят до конца.
		// Так как я не контролирую гоурутины стейджей, они завершатся, когда через них пройдут все значения очереди
		require.Len(t, result, 2)
		// require.Less(t, int64(elapsed), int64(abortDur)+int64(fault))
		// Некорректная формула. Такты по количеству стейджей отработают в любом случае.
		// И отработают такты по количеству просочившихся за время abortDur значений.
		require.Less(t, int64(elapsed),
			int64(sleepPerStage)*int64(len(stages)+int(int64(abortDur)/int64(sleepPerStage))-1)+int64(fault))
	})
}
